huge reorganization of modules, phew
