+++ setup.py
@@ -95,4 +95,4 @@ if __name__ == "__main__":
               'Programming Language :: Python :: 3',
           ],
           entry_points = {'console_scripts': [ 'shub=singularity.cli:main',
-                                               'sreg=singularity.registry.client:main' ] })
+                                               'sreg=singularity.hub.registry:main' ] })
