+++ singularity/analysis/reproduce/__init__.py
@@ -15,6 +15,7 @@ from .hash import (
 
 from .utils import (
     extract_content,
+    delete_image_tar,
     get_memory_tar,
     get_image_tar
 )
