+++ singularity/cli/app/static/img/folder-small.png
