+++ singularity/cli/app/templates/comparison_tree.html
