+++ singularity/cli/app/templates/container_tree.html
