+++ singularity/cli/app/templates/container_tree_circleci.html
