+++ singularity/cli/app/templates/index_circleci.html
