+++ singularity/cli/app/templates/runscript.py
