+++ singularity/cli/app/templates/similarity_scatter.html
