+++ singularity/cli/main/analysis.py
