+++ singularity/cli/main/compare.py
