+++ singularity/cli/main/inspect.py
