+++ singularity/hub/registry/init.py
