+++ singularity/hub/registry/template.py
