+++ singularity/hub/registry/templates/.travis.yml
