+++ singularity/hub/registry/templates/bin/singularity-registry-build
