+++ singularity/hub/registry/templates/bin/singularity-registry-db
