+++ singularity/hub/registry/templates/ci/.travis.yml
