+++ singularity/package/__init__.py
@@ -19,6 +19,7 @@ from .build import (
  
 from .manage import (
     list_package,
+    list_package_families,
     load_package,
     get_packages,
     get_package_base
